# avoid cloning entities when building the response