--- a/crates/apollo-router-core/src/json_ext.rs
+++ b/crates/apollo-router-core/src/json_ext.rs
@@ -24,7 +24,7 @@ pub trait ValueExt {
     /// Deep merge the JSON objects, array and override the values in `&mut self` if they already
     /// exists.
     #[track_caller]
-    fn deep_merge(&mut self, other: &Self);
+    fn deep_merge(&mut self, other: Self);
 
     /// Returns `true` if the values are equal and the objects are ordered the same.
     ///
@@ -123,13 +123,13 @@ impl ValueExt for Value {
         Ok(current)
     }
 
-    fn deep_merge(&mut self, other: &Self) {
+    fn deep_merge(&mut self, other: Self) {
         match (self, other) {
             (Value::Object(a), Value::Object(b)) => {
-                for (key, value) in b.iter() {
+                for (key, value) in b.into_iter() {
                     match a.entry(key) {
                         Entry::Vacant(e) => {
-                            e.insert(value.to_owned());
+                            e.insert(value);
                         }
                         Entry::Occupied(e) => {
                             e.into_mut().deep_merge(value);
@@ -137,15 +137,13 @@ impl ValueExt for Value {
                     }
                 }
             }
-            (Value::Array(a), Value::Array(b)) => {
-                for (index, value) in a.iter_mut().enumerate() {
-                    if let Some(b) = b.get(index) {
-                        value.deep_merge(b);
-                    }
+            (Value::Array(a), Value::Array(mut b)) => {
+                for ((_index, b_value), a_value) in b.drain(..).enumerate().zip(a.iter_mut()) {
+                    a_value.deep_merge(b_value);
                 }
             }
             (a, b) => {
-                *a = b.to_owned();
+                *a = b;
             }
         }
     }
@@ -459,7 +457,7 @@ mod tests {
     #[test]
     fn test_deep_merge() {
         let mut json = json!({"obj":{"arr":[{"prop1":1},{"prop2":2}]}});
-        json.deep_merge(&json!({"obj":{"arr":[{"prop1":2,"prop3":3},{"prop4":4}]}}));
+        json.deep_merge(json!({"obj":{"arr":[{"prop1":2,"prop3":3},{"prop4":4}]}}));
         assert_eq!(
             json,
             json!({"obj":{"arr":[{"prop1":2, "prop3":3},{"prop2":2, "prop4":4}]}})