--- a/crates/apollo-router-core/src/query_planner/model.rs
+++ b/crates/apollo-router-core/src/query_planner/model.rs
@@ -386,43 +386,46 @@ async fn fetch_node<'a>(
             Some(Response {
                 data, mut errors, ..
             }) => {
-                if let Some(entities) = data.get("_entities") {
-                    tracing::trace!(
-                        "Received entities: {}",
-                        serde_json::to_string(entities).unwrap(),
-                    );
-                    if let Some(array) = entities.as_array() {
-                        let mut response = response
-                            .lock()
-                            .instrument(tracing::trace_span!("response_lock_wait"))
-                            .await;
+                if let Value::Object(mut map) = data {
+                    if let Some(entities) = map.remove("_entities") {
+                        tracing::trace!(
+                            "Received entities: {}",
+                            serde_json::to_string(&entities).unwrap(),
+                        );
 
-                        let span = tracing::trace_span!("response_insert");
-                        let _guard = span.enter();
-                        for (i, entity) in array.iter().enumerate() {
-                            response.insert_data(
-                                &current_dir.join(Path::from(i.to_string())),
-                                entity,
-                            )?;
+                        if let Value::Array(mut array) = entities {
+                            let mut response = response
+                                .lock()
+                                .instrument(tracing::trace_span!("response_lock_wait"))
+                                .await;
+
+                            let span = tracing::trace_span!("response_insert");
+                            let _guard = span.enter();
+                            for (i, entity) in array.drain(..).enumerate() {
+                                response.insert_data(
+                                    &current_dir.join(Path::from(i.to_string())),
+                                    entity,
+                                )?;
+                            }
+
+                            return Ok(());
+                        } else {
+                            return Err(FetchError::ExecutionInvalidContent {
+                                reason: "Received invalid type for key `_entities`!".to_string(),
+                            });
                         }
-
-                        Ok(())
-                    } else {
-                        Err(FetchError::ExecutionInvalidContent {
-                            reason: "Received invalid type for key `_entities`!".to_string(),
-                        })
                     }
-                } else {
-                    let mut response = response
-                        .lock()
-                        .instrument(tracing::trace_span!("response_lock_wait"))
-                        .await;
-
-                    response.append_errors(&mut errors);
-                    Err(FetchError::ExecutionInvalidContent {
-                        reason: "Missing key `_entities`!".to_string(),
-                    })
                 }
+
+                let mut response = response
+                    .lock()
+                    .instrument(tracing::trace_span!("response_lock_wait"))
+                    .await;
+
+                response.append_errors(&mut errors);
+                Err(FetchError::ExecutionInvalidContent {
+                    reason: "Missing key `_entities`!".to_string(),
+                })
             }
             None => Err(FetchError::SubrequestNoResponse {
                 service: service_name.to_string(),
@@ -474,7 +477,7 @@ async fn fetch_node<'a>(
                 let span = tracing::trace_span!("response_insert");
                 let _guard = span.enter();
                 response.append_errors(&mut errors);
-                response.insert_data(current_dir, &data)?;
+                response.insert_data(current_dir, data)?;
 
                 Ok(())
             }