--- a/crates/apollo-router-core/src/response.rs
+++ b/crates/apollo-router-core/src/response.rs
@@ -85,7 +85,7 @@ impl Response {
         ))
     }
 
-    pub fn insert_data(&mut self, path: &Path, value: &Value) -> Result<(), FetchError> {
+    pub fn insert_data(&mut self, path: &Path, value: Value) -> Result<(), FetchError> {
         let nodes =
             self.data
                 .get_at_path_mut(path)
@@ -94,7 +94,7 @@ impl Response {
                 })?;
 
         for node in nodes {
-            node.deep_merge(value);
+            node.deep_merge(&value);
         }
 
         Ok(())
@@ -285,7 +285,7 @@ mod tests {
         let data = json!({
             "name": "cook",
         });
-        response.insert_data(&Path::from("job"), &data).unwrap();
+        response.insert_data(&Path::from("job"), data).unwrap();
         assert_eq!(
             response.data,
             json!({