I see a small perf improvement here. With 10k concurrent clients (I'm working with saturated CPUs to see breaking point behaviours):

query:

```
query  {\products(limit: 5) {  upc  name  price  review { body }  } }
```
## main
854MB res memory

```
Summary:                                                         
  Total:        57.0660 secs                                     
  Slowest:      6.7271 secs                                      
  Fastest:      0.0017 secs                                      
  Average:      0.5240 secs                                      
  Requests/sec: 17523.5611                                       
                                                                 
                                                                 
Response time histogram:                                         
  0.002 [1]     |                                                
  0.674 [729343]|■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  1.347 [38168] |■■                                              
  2.019 [207780]|■■■■■■■■■■■                             
  2.692 [4226]  |                                                
  3.364 [16521] |■                                               
  4.037 [3238]  |                                                
  4.709 [293]   |                                                
  5.382 [423]   |                                                
  6.055 [4]     |                                                
  6.727 [3]     |                                                
                                                                 
                                                                 
Latency distribution:                                            
  10% in 0.0463 secs                                             
  25% in 0.0654 secs                                             
  50% in 0.0934 secs                                             
  75% in 1.2584 secs                                             
  90% in 1.6843 secs                                             
  95% in 1.7897 secs                                             
  99% in 3.2215 secs                                             
```

## this PR

830MB res memory

```
Summary:                                                         
  Total:        55.0270 secs                               
  Slowest:      6.3213 secs                                
  Fastest:      0.0013 secs                                
  Average:      0.5042 secs                                
  Requests/sec: 18172.8988                                 
   
                             
Response time histogram: 
  0.001 [1]     |
  0.633 [717221]|■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  1.265 [47129] |■■■
  1.897 [207444]|■■■■■■■■■■■■
  2.529 [5569]  |
  3.161 [18295] |■
  3.793 [3450]  |
  4.425 [478]   |
  5.057 [398]   |
  5.689 [7]     |
  6.321 [8]     |


Latency distribution:
  10% in 0.0447 secs
  25% in 0.0633 secs
  50% in 0.0909 secs
  75% in 1.2007 secs
  90% in 1.5761 secs
  95% in 1.6906 secs
  99% in 3.0063 secs
```

Less memory usage, less latency, increased rps