@Geal 2021-11-18
friendly ping on this. This is an easy perf gain, though I suspect it will have to go away once we do smarter things with the responses and caching