@Geal 2021-11-22
I tested with passing owned values to deep_merge, got this result on large responses (where allocations have a bigger impact):

```
Summary:                                                  
  Total:        21.2887 secs                              
  Slowest:      4.0329 secs                               
  Fastest:      0.3662 secs                               
  Average:      1.9980 secs                               
  Requests/sec: 23.4866                                   
                                                          
                                                          
Response time histogram:                                  
  0.366 [1]     |                                         
  0.733 [2]     |■                                        
  1.100 [27]    |■■■■■■■■                                 
  1.466 [41]    |■■■■■■■■■■■■                             
  1.833 [131]   |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■  
  2.200 [134]   |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 
  2.566 [97]    |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■            
  2.933 [37]    |■■■■■■■■■■■                              
  3.300 [13]    |■■■■                                     
  3.666 [12]    |■■■■                                     
  4.033 [5]     |■                                        
                                                          
                                                          
Latency distribution:                                     
  10% in 1.2888 secs                                      
  25% in 1.6279 secs                                      
  50% in 1.9919 secs                                      
  75% in 2.3022 secs                                      
  90% in 2.7397 secs                                      
  95% in 3.1178 secs                                      
  99% in 3.7767 secs                                      
                                                          
Details (average, fastest, slowest):                      
  DNS+dialup:   0.0004 secs, 0.3662 secs, 4.0329 secs     
  DNS-lookup:   0.0000 secs, 0.0000 secs, 0.0000 secs     
  req write:    0.0002 secs, 0.0000 secs, 0.0200 secs     
  resp wait:    0.3718 secs, 0.0001 secs, 1.1402 secs     
  resp read:    1.6254 secs, 0.3660 secs, 3.5175 secs     
                                                          
Status code distribution:                                 
  [200] 500 responses                                     
```

compare this with the results in https://github.com/apollographql/router/pull/172, it goes from 19rps to 23 rps and reduces overall latency.

It raises an interesting question though. Considering the code here:
https://github.com/apollographql/router/pull/132/files#diff-5c610be4bd672b816c344e0e9cc3a80d07058e95432b5f6c895aa2ffe1d19ca7R97

Right now I don't think there are cases where we can get multiple nodes for the same path (and replacing with only one node does not fail the tests).
I suspect we would get an issue with variable deduplication though: https://github.com/apollographql/router/issues/87
if we deduplicate variables, we can get a case where  a part of a subgraph has to be inserted at multiple points in the response